Factor out VM creation to remember function
